move pkgs around
